+++ gelcoverage/constants.py
@@ -1,7 +1,7 @@
 
-'''
+"""
 This file contains constants that are used in the output JSON as keys.
-'''
+"""
 
 RMSD = "rmsd"
 AVERAGE = "avg"
@@ -41,10 +41,8 @@ UNION_TRANSCRIPT = "union_tr"
 GENES = "genes"
 
 
-AUTOSOME_IDS = set([
+AUTOSOME_IDS = {
     "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20",
-    "21", "22",
-    "chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9", "chr10", "chr11", "chr12", "chr13",
-    "chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21", "chr22"
-])
-
+    "21", "22", "chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9", "chr10", "chr11", "chr12",
+    "chr13", "chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21", "chr22"
+}