+++ gelcoverage/tools/bigwig_reader.py
@@ -3,7 +3,9 @@ import logging
 
 
 class UncoveredIntervalException(Exception):
+
     pass
+
 
 class BigWigReader:
 
@@ -17,13 +19,13 @@ class BigWigReader:
         self.reported_unexisting_chr = []
         logging.debug("BigWig reader created!")
 
-    def read_bigwig_coverages(self, chromosome, start, end, strict = True):
+    def read_bigwig_coverages(self, chromosome, start, end, strict=True):
         """
         Reads the coverage values in a region from the bigwig file
         :param chromosome: the region chromosome
         :param start: the start position
         :param end: the end position
-        :strict: when true if a position not present in the bigwig is queried it will raise an error
+        :param strict: when true if a position not present in the bigwig is queried it will raise an error
         :return: the sequence of coverages (one integer per position)
         """
         logging.debug("Queries bigwig for %s:%s-%s in %s mode" %
@@ -34,9 +36,9 @@ class BigWigReader:
         if strict:
             try:
                 coverages = self.reader.values(chromosome, start, end)
-            except RuntimeError, e:
+            except RuntimeError:
                 # When the queried interval is not present in the bigwig file it returns all 0s coverage and logs it
-                logging.warn("Missing interval in bigwig %s:%s-%s" % (chromosome, start, end))
+                logging.debug("Missing interval in bigwig %s:%s-%s" % (chromosome, start, end))
                 raise UncoveredIntervalException()
         else:
             # By using the function intervals we make sure that we are not querying coordinates not present in the
@@ -44,9 +46,9 @@ class BigWigReader:
             try:
                 intervals = self.reader.intervals(chromosome, start, end)
                 coverages = [coverage for _, _, coverage in intervals]
-            except RuntimeError, e:
+            except RuntimeError:
                 if chromosome not in self.reported_unexisting_chr:
-                    logging.warn("Missing chromosome in bigwig %s" % (chromosome))
+                    logging.debug("Missing chromosome in bigwig %s" % chromosome)
                     self.reported_unexisting_chr.append(chromosome)
                 coverages = [0] * (end - start)
         return coverages
@@ -54,10 +56,7 @@ class BigWigReader:
     def get_chromosome_lengths(self):
         """
         get chromosome lengths from header of bigWig file
-
-        :param bw: pyBigWig file object
-        :param format: specify "dict" to return a dict instead of a bed recognisable format
         :return: list of chromosomes and start and end positions
         """
         logging.debug("Getting chromosomes and their lenght from BigWig header")
-        return {chromosome : [(0, length)] for (chromosome, length) in self.reader.chroms().iteritems()}
+        return {chromosome: [(0, length)] for (chromosome, length) in self.reader.chroms().iteritems()}