+++ gelcoverage/tools/cellbase_helper.py
@@ -4,13 +4,14 @@ from pycellbase.cbconfig import ConfigClient
 import logging
 
 import gelcoverage.stats.sequence_stats as sequence_stats
+import gelcoverage.tools.backoff_retrier as backoff_retrier
 
 
 chromosome_mapping = {
-    "hsapiens" : {
+    "hsapiens": {
         "grch38": {
-            "1":"chr1",
-            "2":"chr2",
+            "1": "chr1",
+            "2": "chr2",
             "3": "chr3",
             "4": "chr4",
             "5": "chr5",
@@ -36,8 +37,8 @@ chromosome_mapping = {
             "MT": "chrM"
         },
         "grch37": {
-            "1":"chr1",
-            "2":"chr2",
+            "1": "chr1",
+            "2": "chr2",
             "3": "chr3",
             "4": "chr4",
             "5": "chr5",
@@ -68,7 +69,7 @@ chromosome_mapping = {
 
 class CellbaseHelper:
 
-    def __init__(self, species, version, assembly, host, filter_flags, filter_biotypes):
+    def __init__(self, species, version, assembly, host, retries, filter_flags, filter_biotypes):
         """
         Initializes the CellBase helper.
         :param species: The species
@@ -83,6 +84,7 @@ class CellbaseHelper:
         self.filter_biotypes = filter_biotypes
         self.assembly = assembly
         self.species = species
+        self.retries = retries
         # Builds JSON configuration
         json_config = {
             "species": species,
@@ -97,6 +99,8 @@ class CellbaseHelper:
         config = ConfigClient(json_config)
         self.cellbase_client = CellBaseClient(config)
         self.cellbase_gene_client = self.cellbase_client.get_gene_client()
+        # Wraps the CB search call into the truncated binary backoff implementation
+        self.cellbase_search = backoff_retrier.wrapper(self.cellbase_gene_client.search, self.retries)
         # Loads chromosome mapping for this specific reference
         self.chromosome_mapping = chromosome_mapping[self.species.lower()][self.assembly.lower()]
 
@@ -117,56 +121,57 @@ class CellbaseHelper:
         """
         return sum([y["annotationFlags"] if "annotationFlags" in y else [] for y in transcripts], [])
 
-    def get_all_gene_names(self, filter=True):
+    def get_all_gene_names(self, _filter=True):
         """
         Gets all existing HGNC gene names
-        :param filter: flag indicating if filtering should be applied
+        :param _filter: flag indicating if filtering should be applied
         :return: list of HGNC gene names
         """
         logging.debug("Getting gene list from CellBase...")
-        cellbase_result = self.cellbase_gene_client.search(
-            assembly=self.assembly,
-            include=",".join(["name", "transcripts.annotationFlags"]),
-            **{"transcripts.biotype": ",".join(self.filter_biotypes) if filter else ""}
-        )
+        # calls to CB
+        cellbase_result = self.cellbase_search(
+                    assembly=self.assembly,
+                    include=",".join(["name", "transcripts.annotationFlags"]),
+                    **{"transcripts.biotype": ",".join(self.filter_biotypes) if _filter else ""}
+                )
         gene_list = [x["name"] for x in cellbase_result[0]["result"]
                      if self.__is_any_flag_included(CellbaseHelper.__get_all_flags_for_gene(x["transcripts"]))]
         logging.debug("Gene list obtained from CellBase of %s genes" % str(len(gene_list)))
         return gene_list
 
-    def __get_gene_info(self, gene_list, filter= True):
+    def __get_gene_info(self, gene_list, _filter=True):
         """
         For a list of HGNC gene names queries CellBase for all information about transcripts and
         their corresponding exons, including the exonic sequence.
         :param gene_list: the list of HGNC gene names
-        :param filter: flag indicating whether to filter by biotypes
+        :param _filter: flag indicating whether to filter by biotypes
         :return: the data structure returned by CellBase
         """
-        cellbase_genes = self.cellbase_gene_client.search(
+        # calls to CB
+        cellbase_genes = self.cellbase_search(
             name=gene_list,
             assembly=self.assembly,
-            include=["name", "chromosome", "transcripts.exons.start",
-                              "transcripts.exons.exonNumber",
-                              "transcripts.id,transcripts.strand",
-                              "transcripts.exons.end", "transcripts.exons.sequence",
-                              "exonNumber", "transcripts.annotationFlags"],
-            **{"transcripts.biotype": self.filter_biotypes if filter else []}
+            include=["name", "chromosome", "transcripts.exons.start", "transcripts.exons.exonNumber",
+                     "transcripts.id,transcripts.strand", "transcripts.exons.end", "transcripts.exons.sequence",
+                     "exonNumber", "transcripts.annotationFlags"],
+            **{"transcripts.biotype": self.filter_biotypes if _filter else []}
         )
         # TODO: check for errors and empty results
         # TODO: unit test
         return cellbase_genes
 
-    def make_exons_bed(self, gene_list, filter=True, has_chr_prefix=False):
+    def make_exons_bed(self, gene_list, _filter=True, has_chr_prefix=False):
         """
         Gets all exons from cellbase and makes a bed - also calculates gc content, returns a valid bed with gc in the
         score column
         :param gene_list The list of genes to be analysed
-        :param filter: flag indicating if filtering should be applied
+        :param _filter: flag indicating if filtering should be applied
+        :param has_chr_prefix: flag indicating if the chromosomes have the chr prefix or not
         :return: pybedtools object
         """
         logging.info("Building gene annotations bed file from CellBase...")
         if gene_list is None or len(gene_list) == 0:
-            raise SystemError("Input gene list is not correct")
+            raise SystemError("Input gene list is empty")
         number_genes = len(gene_list)
 
         # TODO: Verify that all genes in the list are present in the reference
@@ -187,7 +192,7 @@ class CellbaseHelper:
                 for transcript in gene["transcripts"]:
                     filtered_out = "annotationFlags" not in transcript or \
                                    not self.__is_any_flag_included(transcript["annotationFlags"])
-                    if filter and filtered_out:
+                    if _filter and filtered_out:
                         # We ignore transcripts not flagged as any of a set of flags in the config file
                         continue
                     txid = transcript["id"]
@@ -208,8 +213,8 @@ class CellbaseHelper:
         logging.info("Gene annotations bed file built!")
         return bed
 
-    def get_gene_list_from_transcripts (self, transcripts_list):
-        # TODO:
+    def get_gene_list_from_transcripts(self, transcripts_list):
+        # TODO: implement it
         # gene_list = []
         # return gene_list
         raise NotImplemented